registry_service/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
anyio==3.6.1; python_full_version >= '3.6.2'
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3; python_version >= '3.7'
cryptography==37.0.4
dnspython==2.2.1; python_version >= '3.6' and python_version < '4'
ecdsa==0.18.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
fastapi==0.79.0
h11==0.13.0; python_version >= '3.6'
httptools==0.4.0
idna==3.3; python_version >= '3.5'
joblib==1.1.0; python_version >= '3.6'
numpy==1.23.1
passlib[bcrypt]==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1; python_full_version >= '3.6.1'
pymongo[srv]==4.2.0
python-dotenv==0.20.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
pyyaml==6.0
rsa==4.9; python_version >= '3.6' and python_version < '4'
scikit-learn==1.1.1
scipy==1.8.1; python_version < '3.11' and python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
starlette==0.19.1; python_version >= '3.6'
threadpoolctl==3.1.0; python_version >= '3.6'
typing-extensions==4.3.0; python_version < '3.10'
uvicorn[standard]==0.18.2
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
